requirements.txt
altair==5.4.1
annotated-types==0.7.0
anyio==4.6.2.post1
assemblyai==0.34.0
asttokens==2.4.1
attrs==24.2.0
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
decorator==5.1.1
elevenlabs==0.3.0b0
eval_type_backport==0.2.0
executing==2.1.0
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
ipython==8.29.0
jedi==0.19.1
Jinja2==3.1.4
jsonpath-python==1.0.6
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mistralai==1.1.0
mpv==1.0.7
mypy-extensions==1.0.0
narwhals==1.10.0
numpy==2.1.2
packaging==24.1
pandas==2.2.3
parso==0.8.4
pillow==10.4.0
prompt_toolkit==3.0.48
protobuf==5.28.3
pure_eval==0.2.3
pyarrow==17.0.0
PyAudio==0.2.14
pydantic==2.9.2
pydantic_core==2.23.4
pydeck==0.9.1
Pygments==2.18.0
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.2
referencing==0.35.1
requests==2.32.3
rich==13.9.3
rpds-py==0.20.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
SpeechRecognition==3.11.0
stack-data==0.6.3
streamlit==1.39.0
tenacity==9.0.0
toml==0.10.2
tornado==6.4.1
traitlets==5.14.3
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
watchdog==5.0.3
wcwidth==0.2.13
websockets==13.1
